Set port count for new SATA controllers to 1
The default port count for new SATA controllers is 30. This increases the
boot time for CentOS (and possibly other) VMs while each port is probed for a
disk. Currently veewee only ever adds one disk to port 0 of the SATA
controller and Vagrant doesn't support the automatic attachment of disks to
a box during `vagrant up`, so it should be safe to reduce this to 1.

This may need to be incremented if #86 is implemented or configurable if
feature 334 is implemented in Vagrant.