Allow users to set VEEWEE_DIR env var to override
Currently Veewee expects a directory structure to exist under
the present working directory (PWD), but this is not how everyone
would like to structure their project in their project root directory.

This is the simplest fix for those of us that wish to have our basebox
definitions under a subdirectory by defining VEEWEE_DIR as an environment
variable.