Platform check error : https://github.com/whren/veewee/issues/1
In fact the identification of windows platform is not needeed as i used this to specialize widnows behavior.
The fix i've already made on the split for the 'Default Machine' VBoxManage system property should work on all os because i just said to split function to limit its split to 2 elements : the parameter name and its value.
So here i remove the windows platform check and it'll work.


Signed-off-by: Whren <whren@free.fr>