Remove postinstall files matching exclude regex
Previous to this commit, postinstall files matching the provided exclude
regex were not removed from the postinstall list, but rather were just
prepended with '_'.  This caused an error since the script '_<original
filename>' doesn't exist in the definition.

This commit removes the files from the list of postinstall files if the
filename matches the provided regex