Properly determine if a volume or instance exists
Preious to this commit, volumes were determined to exist or not by
calling the Volumes#all method the :name parameter to filter only the
volumes that matched the value of :name.  This was not properly
finding the volumes.  This commit uses the #list_volumes method and
calls #find on the array returned.

Likewise for instances