Make IP retrieval more reliable
Previous to this commit, the public IP returned from ruby-libvirt may
not be an array. The code assumes it's an array and always grabs the
first item.  I'm not sure if it's ever an array, but to be safe this
commit ensures what's returned from #public_ip_address is an array, then
the first item is selected.