Correctly modify /etc/sudeors to allow passwordsless sudo.
The sed expression in the previous postinstall.sh file modifications
I had made used 4 spaces instead of a single '\t' which prevented
the required substition in the /etc/sudoers file which prevented the
vagrant user from being able to sudo without provide a password. The
sed expression now correctly substitutes the directives for the
sudo group so passwordless sudo is enabled.