Skip copy/chmod of disabled post install files
Currently, if you disable a postinstall_script by adding
an underscore, veewee will try to copy that file over with
and underscore in the name. Since the underscore is being
used to disable a post install script and not point to a
specific file, we should just skip copying disabled postinstall
scripts over since they will not be used.