Update installation.md
I've changed (would like to change) the ruby versions used in the documentation to reflect a more current ruby version.  1.9.2 refuses to compile at all on the latest version of OS X (Yosemite).  I've proposed 2.0.0-p598 because a) I've installed Veewee on an Ubuntu 12.04 machine and my OS X machine within the last 24 hours, and successfully built a box without any noticed ruby errors,  and b) I received advice on https://github.com/jedi4ever/veewee/issues/774 that seemed confident that 2.0.0 was the correct version supported by Veewee.

I've also stuck a h5 at the top of the doc to assert that this is the supported version (though have not included a patch number).

(I think it would also be good if the notes regarding v1.8.7 were removed completely but I don't have the information about who's using which versions of ruby).